samd21, samd51: move USB-CDC code
